WSTEAM1-493: Refactor
