+++ src/app/components/MediaPlayer/fixture.ts
